requirements.txt
gradio==5.24.0
hydra-core==1.3.2
mlflow==3.5.1
numpy==2.1.1
optuna==3.6.1
pandas==2.2.2
plotly==6.3.1
ray==2.49.0
scikit-learn==1.5.1
torch==2.4.0
torchinfo==1.8.0
uv==0.6.6